feat: move logic to change settings to the preload
