+++ packages/creator-hub/preload/src/modules/settings.ts
@@ -1,4 +1,5 @@
 import path from 'path';
+import fs from 'fs/promises';
 import { ipcRenderer, type IpcRendererEvent } from 'electron';
 import {
   DEPENDENCY_UPDATE_STRATEGY,
@@ -107,5 +108,53 @@ export async function getEditors() {
 }
 
 export async function setDefaultEditor(editorPath: string) {
-  return invoke('defaultEditor.setDefaultEditor', editorPath);
+  let name: string;
+  if (process.platform === 'darwin') {
+    const fileName = path.basename(editorPath);
+    if (!fileName.endsWith('.app')) {
+      throw new Error('invalid_app_extension');
+    }
+    const macosPath = path.join(editorPath, 'Contents', 'MacOS');
+    try {
+      await fs.stat(macosPath);
+    } catch {
+      throw new Error('invalid_app_bundle');
+    }
+    name = fileName.replace('.app', '');
+  } else {
+    const fileName = path.basename(editorPath);
+    if (!fileName.endsWith('.exe')) {
+      throw new Error('invalid_exe_file');
+    }
+    name = fileName.replace('.exe', '') || 'Custom Editor';
+  }
+
+  await setConfig(config => {
+    if (!config.editors) {
+      config.editors = [];
+    }
+
+    const existingIndex = config.editors.findIndex(e => e.name === name);
+
+    config.editors.forEach(editor => {
+      editor.isDefault = false;
+    });
+
+    if (existingIndex >= 0) {
+      config.editors[existingIndex] = {
+        ...config.editors[existingIndex],
+        path: editorPath,
+        isDefault: true,
+      };
+    } else {
+      config.editors.push({
+        name,
+        path: editorPath,
+        isDefault: true,
+      });
+    }
+  });
+
+  const updatedConfig = await getConfig();
+  return updatedConfig.editors;
 }
